requirements.txt
flask
requests
beautifulsoup4
nltk
scikit-learn
python-dotenv
newsapi-python
